re-structure tests; add allgather test
