+++ tests/collective_ops/test_sendrecv.py
@@ -1,0 +1,106 @@
+import pytest
+
+from mpi4py import MPI
+
+import jax
+import jax.numpy as jnp
+
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+size = comm.Get_size()
+
+
+@pytest.mark.skipif(size < 2 or rank > 1, reason="Runs only on rank 0 and 1")
+def test_sendrecv():
+    from mpi4jax import Sendrecv
+
+    arr = jnp.ones((3, 2)) * rank
+    _arr = arr.copy()
+
+    other = 1 - rank
+
+    res, token = Sendrecv(arr, arr, source=other, dest=other)
+
+    assert jnp.array_equal(res, jnp.ones_like(arr) * other)
+    assert jnp.array_equal(_arr, arr)
+
+
+@pytest.mark.skipif(size < 2 or rank > 1, reason="Runs only on rank 0 and 1")
+def test_sendrecv_status():
+    from mpi4jax import Sendrecv
+
+    arr = jnp.ones((3, 2)) * rank
+    _arr = arr.copy()
+
+    other = 1 - rank
+
+    status = MPI.Status()
+    res, token = Sendrecv(arr, arr, source=other, dest=other, status=status)
+
+    assert jnp.array_equal(res, jnp.ones_like(arr) * other)
+    assert jnp.array_equal(_arr, arr)
+    assert status.Get_source() == other
+
+
+@pytest.mark.skipif(size < 2 or rank > 1, reason="Runs only on rank 0 and 1")
+def test_sendrecv_status_jit():
+    from mpi4jax import Sendrecv
+
+    arr = jnp.ones((3, 2)) * rank
+    _arr = arr.copy()
+
+    other = 1 - rank
+
+    status = MPI.Status()
+    res = jax.jit(
+        lambda x, y: Sendrecv(x, y, source=other, dest=other, status=status)[0]
+    )(arr, arr)
+
+    assert jnp.array_equal(res, jnp.ones_like(arr) * other)
+    assert jnp.array_equal(_arr, arr)
+    assert status.Get_source() == other
+
+
+@pytest.mark.skipif(size < 2 or rank > 1, reason="Runs only on rank 0 and 1")
+def test_sendrecv_scalar():
+    from mpi4jax import Sendrecv
+
+    arr = 1 * rank
+    _arr = arr
+
+    other = 1 - rank
+
+    res, token = Sendrecv(arr, arr, source=other, dest=other)
+
+    assert jnp.array_equal(res, jnp.ones_like(arr) * other)
+    assert jnp.array_equal(_arr, arr)
+
+
+@pytest.mark.skipif(size < 2 or rank > 1, reason="Runs only on rank 0 and 1")
+def test_sendrecv_jit():
+    from mpi4jax import Sendrecv
+
+    arr = jnp.ones((3, 2)) * rank
+    _arr = arr.copy()
+
+    other = 1 - rank
+
+    res = jax.jit(lambda x, y: Sendrecv(x, y, source=other, dest=other)[0])(arr, arr)
+
+    assert jnp.array_equal(res, jnp.ones_like(arr) * other)
+    assert jnp.array_equal(_arr, arr)
+
+
+@pytest.mark.skipif(size < 2 or rank > 1, reason="Runs only on rank 0 and 1")
+def test_sendrecv_scalar_jit():
+    from mpi4jax import Sendrecv
+
+    arr = 1 * rank
+    _arr = arr
+
+    other = 1 - rank
+
+    res = jax.jit(lambda x, y: Sendrecv(x, y, source=other, dest=other)[0])(arr, arr)
+
+    assert jnp.array_equal(res, jnp.ones_like(arr) * other)
+    assert jnp.array_equal(_arr, arr)
